+++ frontend/src/metabase/dashboard/containers/DashboardHeader.jsx
@@ -33,7 +33,7 @@ import {
 import { hasDatabaseActionsEnabled } from "metabase/dashboard/utils";
 import { saveDashboardPdf } from "metabase/visualizations/lib/save-dashboard-pdf";
 
-import Header from "../components/DashboardHeader";
+import DashboardHeaderView from "../components/DashboardHeaderView";
 import { SIDEBAR_NAME } from "../constants";
 import {
   DashboardHeaderButton,
@@ -445,7 +445,7 @@ class DashboardHeader extends Component {
     const hasLastEditInfo = dashboard["last-edit-info"] != null;
 
     return (
-      <Header
+      <DashboardHeaderView
         headerClassName="wrapper"
         objectType="dashboard"
         analyticsContext="Dashboard"
